#20: Generate graphql by run command line and run configuration
