+++ backend/build.gradle.kts
@@ -27,6 +27,7 @@ repositories {
 }
 
 dependencies {
+    implementation("com.github.ajalt.clikt:clikt:3.0.1")
     implementation("io.javalin:javalin:3.11.0")
     implementation("com.google.code.gson", "gson", "2.8.6")
     implementation("org.slf4j", "slf4j-simple", "1.7.30")
@@ -44,7 +45,7 @@ dependencies {
 
     // Use the Kotlin JUnit integration.
     testImplementation("org.jetbrains.kotlin:kotlin-test-junit")
-    
+
     implementation("org.jetbrains.exposed", "exposed-core", exposed_version)
     implementation("org.jetbrains.exposed", "exposed-dao", exposed_version)
     implementation("org.jetbrains.exposed", "exposed-jdbc", exposed_version)
@@ -53,7 +54,7 @@ dependencies {
 
 application {
     // Define the main class for the application.
-    mainClassName = "xyz.elitese.ehrenamtskarte.webservice.EntrypointKt"
+    mainClassName = "xyz.elitese.ehrenamtskarte.EntryPointKt"
 }
 
 tasks.withType<JavaExec>().configureEach {
