+++ frontend/schema.graphql
@@ -1,0 +1,82 @@
+schema {
+  query: Query
+}
+
+"Directs the executor to include this field or fragment only when the `if` argument is true"
+directive @include(
+    "Included when true."
+    if: Boolean!
+  ) on FIELD | FRAGMENT_SPREAD | INLINE_FRAGMENT
+
+"Directs the executor to skip this field or fragment when the `if`'argument is true."
+directive @skip(
+    "Skipped when true."
+    if: Boolean!
+  ) on FIELD | FRAGMENT_SPREAD | INLINE_FRAGMENT
+
+"Marks the field or enum value as deprecated"
+directive @deprecated(
+    "The reason for the deprecation"
+    reason: String = "No longer supported"
+  ) on FIELD_DEFINITION | ENUM_VALUE
+
+"Exposes a URL that specifies the behaviour of this scalar."
+directive @specifiedBy(
+    "The URL that specifies the behaviour of this scalar."
+    url: String!
+  ) on SCALAR
+
+type AcceptingStore {
+  category: Category!
+  categoryId: Long!
+  contact: Contact!
+  contactId: Long!
+  id: Long!
+  name: String!
+  phyiscalStore: PhysicalStore
+}
+
+type Address {
+  coordinates: Coordinates!
+  houseNumber: String!
+  location: String!
+  postalCode: String!
+  state: String!
+  street: String!
+}
+
+type Category {
+  iconUrl: String!
+  id: Long!
+  name: String!
+}
+
+type Contact {
+  email: String!
+  id: Long!
+  telephone: String!
+  website: String!
+}
+
+type Coordinates {
+  latitude: Float!
+  longitude: Float!
+}
+
+type PhysicalStore {
+  address: Address!
+  id: Long!
+}
+
+type Query {
+  acceptingStoreById(params: ParamsInput!): [AcceptingStore]!
+  "Return list of all accepting stores."
+  acceptingStores: [AcceptingStore!]!
+}
+
+"Long type"
+scalar Long
+
+input ParamsInput {
+  ids: [Long!]!
+}
